Add labels and annotations to Elyra pods (#83)
Properly label Elyra pods to enable cluster administrators 
to identify which pods are associated with Elyra pipelines (in general)
as a well as enabling correlation with the artifact associated with the pod

Labels metadata:
elyra-node-type: static string identifying the node type as notebook-script, 
which covers file-based Elyra container ops for notebooks and Python scripts

elyra-pipeline-name: the pipeline's experiment name, 
e.g. my-labeled-pipeline3-0208165033 which is not the same as the pipeline name

elyra-node-name: the pipeline node name (see also elyra-ai/elyra#971)

Annotations metadata:
elyra-node-file-name: the notebook or Python script name, e.g. notebook1.ipynb